Moving Reducers and views into appropriate folders
